Improved ElementSlideAnimator view handler
Renamed class and file to be more descriptive and consistent.
Created extra methods for showing and hiding.
Added comment why null value is default instead of true/false.
Renamed property to reflect type.
