+++ app/js/main.js
@@ -11,7 +11,7 @@ module.exports = {
 	StoreUpdates: require( './lib/store_update_handling' ),
 	View: {
 		createClearAmountHandler: require( './lib/view_handler/clear_amount' ).createHandler,
-		createElementSliderHandler: require( './lib/view_handler/element_slide' ).createHandler,
+		createElementSlideAnimationHandler: require( './lib/view_handler/element_slide_animator' ).createHandler,
 		createErrorBoxHandler:  require( './lib/view_handler/error_box' ).createHandler,
 		createFormPageVisibilityHandler: require( './lib/view_handler/form_page_visibility' ).createHandler
 	},
