+++ app/templates/DonationForm.html.twig
@@ -79,32 +79,32 @@
                         stateKey: 'validationMessages'
                     },
                     {
-                        viewHandler: WMDE.View.createElementSliderHandler( $( '.periode-2-list' ), /^(1|3|6|12)$/ ),
+                        viewHandler: WMDE.View.createElementSlideAnimationHandler( $( '.periode-2-list' ), /^(1|3|6|12)$/ ),
                         stateKey: 'formContent.paymentPeriodInMonths'
                     },
                     {
-                        viewHandler: WMDE.View.createElementSliderHandler( $( '#bank-data' ), 'BEZ' ),
+                        viewHandler: WMDE.View.createElementSlideAnimationHandler( $( '#bank-data' ), 'BEZ' ),
                         stateKey: 'formContent.paymentType'
                     },
                     {
-                        viewHandler: WMDE.View.createElementSliderHandler( $( '.slide-sepa' ), 'sepa' ),
+                        viewHandler: WMDE.View.createElementSlideAnimationHandler( $( '.slide-sepa' ), 'sepa' ),
                         stateKey: 'formContent.debitType'
                     },
                     {
-                        viewHandler: WMDE.View.createElementSliderHandler( $( '.slide-non-sepa' ), 'non-sepa' ),
+                        viewHandler: WMDE.View.createElementSlideAnimationHandler( $( '.slide-non-sepa' ), 'non-sepa' ),
                         stateKey: 'formContent.debitType'
                     },
                     // Show only the right data fields for personal data
                     {
-                        viewHandler: WMDE.View.createElementSliderHandler( $( '.personal-data-person' ), 'person' ),
+                        viewHandler: WMDE.View.createElementSlideAnimationHandler( $( '.personal-data-person' ), 'person' ),
                         stateKey: 'formContent.addressType'
                     },
                     {
-                        viewHandler: WMDE.View.createElementSliderHandler( $( '.personal-data-company' ), 'firma' ),
+                        viewHandler: WMDE.View.createElementSlideAnimationHandler( $( '.personal-data-company' ), 'firma' ),
                         stateKey: 'formContent.addressType'
                     },
                     {
-                        viewHandler: WMDE.View.createElementSliderHandler( $( '.personal-data-full' ), /firma|person/ ),
+                        viewHandler: WMDE.View.createElementSlideAnimationHandler( $( '.personal-data-full' ), /firma|person/ ),
                         stateKey: 'formContent.addressType'
                     }
                 ],
